chore: bump anthropic SDK 0.23.1 -> 0.109.1; fork claude prediction tools to _v1
Lockstep bump of the anthropic SDK across every pinning package. Tools whose
behaviour changes under the bump are forked to new _v1 packages (originals
removed so the single shared env stays on one anthropic version);
output-preserving tools get an in-place pin bump only.

- pyproject + 11 component.yaml pins -> anthropic==0.109.1
- Forked (model claude-4-sonnet-20250514 -> claude-sonnet-4-6; every registered
  tool name suffixed -v1; originals deleted; live messages.count_tokens now
  used in place of the tiktoken fallback):
    prediction_request            -> prediction_request_v1
    prediction_request_rag        -> prediction_request_rag_v1
    prediction_url_cot            -> prediction_url_cot_v1
    prediction_request_reasoning  -> prediction_request_reasoning_v1
- In-place pin bump (output-preserving): resolve_market_reasoning, prepare_tx,
  resolve_market, prediction_langchain, prediction_request_sme,
  factual_research_v3, superforcaster_polymarket_v3
- tool_lineage.json (+10 entries), benchmark/tools.py TOOL_REGISTRY,
  packages.json, agent aea-config customs list updated; CIDs re-locked

Phase 0 resolvability confirmed: anthropic 0.109.1 resolves against the frozen
open-autonomy 0.21.23 / httpx 0.25.2 contract with no open-autonomy bump.
Not registered in tournament_tools.json by request.
