+++ packages/napthaai/customs/prediction_request_rag_v1/component.yaml
@@ -1,4 +1,4 @@
-name: prediction_request_rag
+name: prediction_request_rag_v1
 author: napthaai
 version: 0.1.0
 type: custom
@@ -7,11 +7,11 @@ license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: bafybeibt7f7crtwvmkg7spy3jhscmlqltvyblzp32g6gj44v7tlo5lycuq
-  prediction_request_rag.py: bafybeielwbai4tejjkdarsgdavurswv4mat62e6lcdrkctwkxnbkq3bae4
+  prediction_request_rag_v1.py: bafybeiatvqlbp5osiygvskuy4cgq2jeuli2zvlro5blgaosj6cyzvlemsu
   tests/__init__.py: bafybeifcgilmgfwx7kaap67cfnuokrhfrabi6bnvqiudyzgf2idu64dvxq
-  tests/test_prediction_request_rag.py: bafybeihfaopk37fqncboshyvq24f4vwv3a3rvpdsj7ttyb5eacokxhcx34
+  tests/test_prediction_request_rag_v1.py: bafybeiggrgqpojvgfbcvbz7g7ogjfjfwsj2u27zqn24xzfnyjcvflkhtg4
 fingerprint_ignore_patterns: []
-entry_point: prediction_request_rag.py
+entry_point: prediction_request_rag_v1.py
 callable: run
 params:
   default_model: gpt-4.1-2025-04-14
@@ -31,7 +31,7 @@ dependencies:
   lxml_html_clean:
     version: ==0.4.4
   anthropic:
-    version: ==0.23.1
+    version: ==0.109.1
   tiktoken:
     version: ==0.12.0
   pydantic:
