+++ packages/napthaai/customs/prediction_request_reasoning_v1/component.yaml
@@ -1,4 +1,4 @@
-name: prediction_request_reasoning
+name: prediction_request_reasoning_v1
 author: napthaai
 version: 0.1.0
 type: custom
@@ -8,11 +8,11 @@ license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: bafybeib36ew6vbztldut5xayk5553rylrq7yv4cpqyhwc5ktvd4cx67vwu
-  prediction_request_reasoning.py: bafybeib3422icio62o6dbun2qfljdkdsruurtk3yt7rp4w6e24tmae2cmi
+  prediction_request_reasoning_v1.py: bafybeifqplzx5g5vubiq6ft5rzeb44fpwmk4otniedqardscy6muue3zvq
   tests/__init__.py: bafybeieu3toasuyaehkqounrtylk5bjer7qgvnt6ccjcsi6jlfig7wfrka
-  tests/test_prediction_request_reasoning.py: bafybeiclefskba2fa4rt4tohvxiepm3uevqul5zwqo6zk6v3e5mpbrysaq
+  tests/test_prediction_request_reasoning_v1.py: bafybeifayzjezn2iaakrfujjircp7my5jwy3qlalbepeactikakabau2mu
 fingerprint_ignore_patterns: []
-entry_point: prediction_request_reasoning.py
+entry_point: prediction_request_reasoning_v1.py
 callable: run
 params:
   default_model: gpt-4.1-2025-04-14
@@ -46,4 +46,4 @@ dependencies:
   docstring-parser:
     version: ==0.17.0
   anthropic:
-    version: ==0.23.1
+    version: ==0.109.1
