+++ packages/napthaai/customs/prediction_url_cot_v1/__init__.py
