+++ packages/napthaai/customs/resolve_market_reasoning/component.yaml
@@ -45,4 +45,4 @@ dependencies:
   numpy:
     version: '>=1.19.0'
   anthropic:
-    version: ==0.23.1
+    version: ==0.109.1
