+++ packages/packages.json
@@ -2,15 +2,15 @@
     "dev": {
         "custom/dvilela/corcel_request/0.1.0": "bafybeic74xc32orfiffumv3cxe7o4vktzobk6gszjecuhto4or7k5ppi5y",
         "custom/dvilela/gemini_prediction/0.1.0": "bafybeifpoil2wjh6hr7sqoomlilhsyt3ktpy6tjzmisrlopxcjbkr3hhfq",
-        "custom/napthaai/prediction_request_rag/0.1.0": "bafybeicbht6wr4lgcadxedcqsekmm5yechqba2pj3gy6fyvqa6nu5epvpa",
-        "custom/napthaai/prediction_request_reasoning/0.1.0": "bafybeifq3zndol4reyix7aivttvof4xh5adkcmiv7o4bswazgbeym6jqsq",
-        "custom/napthaai/prediction_url_cot/0.1.0": "bafybeibap2akjuqysqdrwvrpac5uwssoes4najrbq6dhg3mvecp3wkxmsy",
-        "custom/napthaai/resolve_market_reasoning/0.1.0": "bafybeibjmyiqm5nnw75y6q7x4vymurg5w74zsko3ewzvarshgq4npxew3y",
-        "custom/nickcom007/prediction_request_sme/0.1.0": "bafybeifssc6ed7itkvqot6qunlpd3prghwsf5dh26kin3xduci3sohulqq",
-        "custom/valory/prediction_langchain/0.1.0": "bafybeifuxvvldtt7o5zmrgdscbqofnkgtmemv5yv5gxzeptnxsiymwvs6a",
-        "custom/valory/prediction_request/0.1.0": "bafybeieqpji2z666rh6ujeldooq23uutpjujvggd3e5dagtjdxjrtty36e",
-        "custom/valory/prepare_tx/0.1.0": "bafybeida74hcfsfgh644o27qlq2lmue5aj7kvjfsbjermwv42euyf525ya",
-        "custom/valory/resolve_market/0.1.0": "bafybeigxgcjsvi2zovoxgz5ra7x7sqz5qfx3p5kxmhgftrr7fxonjmhspi",
+        "custom/napthaai/prediction_request_rag_v1/0.1.0": "bafybeib62kyn5piaup7zhobbgkupkhght4frohhx35enqvsbzqq42uti7e",
+        "custom/napthaai/prediction_request_reasoning_v1/0.1.0": "bafybeih6euompjmuclgrpyuuzzms4rc6y3luht26awan75hcrxwrnemswq",
+        "custom/napthaai/prediction_url_cot_v1/0.1.0": "bafybeig3dhs3wgxdqbqwm3tnxzbgx4nlyzwrnzrhhtmdmi4nomavedk3au",
+        "custom/napthaai/resolve_market_reasoning/0.1.0": "bafybeidji6or6kpmawho64tc7qetinykhrklvv2gmnqfcd6ejmg43j3i7q",
+        "custom/nickcom007/prediction_request_sme/0.1.0": "bafybeifsk4og24t22agcrj7mm6fz3wxwfbdg554yjhc6odpxal5ofg22ti",
+        "custom/valory/prediction_langchain/0.1.0": "bafybeidtixb6gkctowphterybnlxwvb6rrzj2z2wwbcfyalgfb2at3r63q",
+        "custom/valory/prediction_request_v1/0.1.0": "bafybeigc632jofeuvax4c4v52de74hrpsfsyiudobbkk6ryh4iukj4gztq",
+        "custom/valory/prepare_tx/0.1.0": "bafybeidpvkwtn5m5yjp5mr2tn6f52btmlki32syahvkkuvlw3oq5eee3di",
+        "custom/valory/resolve_market/0.1.0": "bafybeiaozazbggoglwrnfyn5v2qebyn4y4jpaxxbdodi3bd5eyieo4a55i",
         "custom/valory/superforcaster/0.1.0": "bafybeidkraklf5c6rkv6g7y7lqj2woy4j5u5dmupjez6mtjhdek3lzbmki",
         "custom/victorpolisetty/dalle_request/0.1.0": "bafybeiadatvfc6opcsmhpxxzmsqpwgbckblciqypx2iwe5uwove6nmki3e",
         "custom/victorpolisetty/gemini_request/0.1.0": "bafybeiamjk5mfycjqkstkzymggako2jt7yjros2zx4l54zyuei2xkj4gqu",
@@ -19,12 +19,12 @@
         "custom/valory/superforcaster_full_search/0.1.0": "bafybeihba6ch6g7gndfwv75dtpfrpnvg4dw6wq6lxexch5h3n3vjafuxsm",
         "custom/valory/superforcaster_calibrated_full_search/0.1.0": "bafybeialuhbtfumvlbdtzimfn7wdacqwccn2ny5cbhi4tkri6iih6pfgxe",
         "custom/valory/superforcaster_polymarket_v1/0.1.0": "bafybeiawfhy3w5aclmxm5zxebwd5l7paseqlgmihwbqfyhcdpn2kccgtrq",
-        "custom/valory/superforcaster_polymarket_v3/0.1.0": "bafybeigwcghcfq4d7zbjixg3d4mqvyo6ci4wp4jfstuyhm5357bhz2aeoy",
+        "custom/valory/superforcaster_polymarket_v3/0.1.0": "bafybeicghoipp43ld56ikew2p3tvwkmzrzm6fdg2y4qjkwrsj77cwioxgy",
         "custom/valory/factual_research_v1/0.1.0": "bafybeicg52wve2ytam3eisks7gumvrswm2v6bkuafburdlmh6d553nrvee",
         "custom/valory/factual_research_v2/0.1.0": "bafybeidgo3jesmvuwa64ooij7mywk7hf7ls5len2sf7cbcjwt3t6r5u6qm",
-        "custom/valory/factual_research_v3/0.1.0": "bafybeih2ciwkf7edrmwlb5knk4nw5dedh7cz2lo6d6ecyehxlt6q7i7i3a",
-        "agent/valory/mech_predict/0.1.0": "bafybeigkwuflnmxza4g2hwvjreohqh5r5zfrbscq4qhe5jyyvn623c6wpq",
-        "service/valory/mech_predict/0.1.0": "bafybeielumgxe6ls36ot4ybgnrlv3bpuy6cxvuymgyvnqfe3fddpbx3ko4"
+        "custom/valory/factual_research_v3/0.1.0": "bafybeigzkzvdmomb3zpcticgaftu2cvo4wssah76gry5y6tjv3ng7eh3de",
+        "agent/valory/mech_predict/0.1.0": "bafybeiav2rt53gmnd4tpmabqfuprl4c52xtkeylynmpzhqmcugz73jlbue",
+        "service/valory/mech_predict/0.1.0": "bafybeigxelbokml7otxnnfxklkax5khsd6376ywdtvpixghvcsw7i562ei"
     },
     "third_party": {
         "protocol/open_aea/signing/1.0.0": "bafybeifsjmldwyki3beqyvdt5lzenrg6wyrqaar5plc5rpnvtc4zlentye",
