+++ packages/valory/customs/factual_research_v3/component.yaml
@@ -26,7 +26,7 @@ dependencies:
   openai:
     version: ==1.93.0
   anthropic:
-    version: ==0.23.1
+    version: ==0.109.1
   tiktoken:
     version: ==0.12.0
   pydantic:
