+++ packages/valory/customs/prediction_langchain/component.yaml
@@ -27,7 +27,7 @@ dependencies:
   httpx:
     version: ==0.25.2
   anthropic:
-    version: ==0.23.1
+    version: ==0.109.1
   google-api-python-client:
     version: ==2.95.0
   typing_extensions:
