+++ packages/valory/customs/prepare_tx/component.yaml
@@ -17,6 +17,6 @@ dependencies:
   httpx:
     version: ==0.25.2
   anthropic:
-    version: ==0.23.1
+    version: ==0.109.1
   google-api-python-client:
     version: ==2.95.0
