+++ packages/valory/customs/superforcaster_polymarket_v3/component.yaml
@@ -27,7 +27,7 @@ dependencies:
   openai:
     version: ==1.93.0
   anthropic:
-    version: ==0.23.1
+    version: ==0.109.1
   httpx:
     version: ==0.25.2
   tiktoken:
