+++ packages/valory/services/mech_predict/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeifogoiwjqlb3nk2o6bbybr4safi4bfv4va7a4rq7ptse6ps7ep4zm
 fingerprint_ignore_patterns: []
-agent: valory/mech_predict:0.1.0:bafybeigkwuflnmxza4g2hwvjreohqh5r5zfrbscq4qhe5jyyvn623c6wpq
+agent: valory/mech_predict:0.1.0:bafybeiav2rt53gmnd4tpmabqfuprl4c52xtkeylynmpzhqmcugz73jlbue
 number_of_agents: 1
 deployment:
   agent:
