+++ pyproject.toml
@@ -36,7 +36,7 @@ dependencies = [
     "googlesearch-python==1.2.3",
     "beautifulsoup4==4.12.2",
     "google-api-python-client==2.95.0",
-    "anthropic==0.23.1",
+    "anthropic==0.109.1",
     "langchain==1.2.10",
     "langchain-core==1.2.17",
     "langchain-tavily==0.2.17",
