+++ tool_lineage.json
@@ -2,6 +2,46 @@
   "version": 1,
   "_doc": "Ledger of tool variants, keyed by tool name. See CLAUDE.md section 'Tool-improvement housekeeping rules' for the canonical schema and update rule.",
   "tools": {
+    "prediction-online-v1": {
+      "parent": "prediction-online",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "prediction-offline-v1": {
+      "parent": "prediction-offline",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "claude-prediction-online-v1": {
+      "parent": "claude-prediction-online",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "claude-prediction-offline-v1": {
+      "parent": "claude-prediction-offline",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "prediction-request-rag-v1": {
+      "parent": "prediction-request-rag",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "prediction-request-rag-claude-v1": {
+      "parent": "prediction-request-rag-claude",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "prediction-url-cot-v1": {
+      "parent": "prediction-url-cot",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "prediction-url-cot-claude-v1": {
+      "parent": "prediction-url-cot-claude",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "prediction-request-reasoning-v1": {
+      "parent": "prediction-request-reasoning",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
+    "prediction-request-reasoning-claude-v1": {
+      "parent": "prediction-request-reasoning-claude",
+      "reason": "anthropic SDK 0.23.1->0.109.1 lockstep bump; dead model claude-4-sonnet-20250514 swapped to claude-sonnet-4-6."
+    },
     "superforcaster-polymarket-v1": {
       "parent": "superforcaster",
       "reason": "Carve out the older uncalibrated superforcaster as a Polymarket-specific tool, leaving the original superforcaster untouched."
